Changing class name from Type to MedicalSpecialization and changing related variables names in the same way
